+++ resources/js/components/IndexField.vue
@@ -1,5 +1,6 @@
 <template>
-  <span>{{ value }}</span>
+  <span v-if="!field.asHtml">{{ value }}</span>
+  <span v-else v-html="value"></span>
 </template>
 
 <script>
@@ -21,14 +22,31 @@ export default {
           .filter(Boolean)
           .map(val => `${this.isOptionGroups ? `[${val.group}] ` : ''}${val.label}`);
 
-        const joinedValues = values.join(', ');
-        if (joinedValues.length <= 40) return joinedValues;
+        const joinedValues = values.join(this.delimiter);
+
+        if (this.valueDisplayLimit >= values.length && this.charDisplayLimit >= joinedValues.length) {
+          return joinedValues;
+        }
 
         return this.__('novaMultiselect.nItemsSelected', { count: String(values.length || '') });
       } else {
         const value = this.field.options.find(opt => String(opt.value) === String(this.field.value));
         return (value && value.label) || '—';
       }
+    },
+
+    delimiter() {
+      return this.field.indexDelimiter ?? ', ';
+    },
+
+    valueDisplayLimit() {
+      return this.field.indexValueDisplayLimit ?? 9999;
+    },
+
+    charDisplayLimit() {
+      // Set char limit to 9999 if we have value limit, but not char limit
+      if (!!this.field.indexValueDisplayLimit && !this.field.indexCharDisplayLimit) return 9999;
+      return this.field.indexCharDisplayLimit ?? 40;
     },
   },
 };