+++ CHANGELOG.md
@@ -5,6 +5,19 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [2.3.0] - 2021-06-16
+
+### Added
+
+- Added three new functions
+  - `indexDelimiter` - Define the delimiter used for index display
+  - `indexValueDisplayLimit` - Define how many values to display at once
+  - `indexCharDisplayLimit` - Set char limit for index display value
+
+### Changed
+
+- Updated packages
+
 ## [2.2.1] - 2021-05-06
 
 ### Removed